requirements.txt
Pillow==10.3.0
pytesseract==0.3.10
Requests==2.32.3
